added pmier suggested changes
